Refactor move generation: Remove move_picker types and enhance legality tests
- Deleted the `types.rs` file which contained definitions for `ScoredMove` and `MovePickerStage`, streamlining the move generation process.
- Added comprehensive tests for move generation completeness, ensuring all legal moves are generated and no duplicates exist.
- Migrated drop restriction tests from MovePicker to position tests, covering various scenarios including pawn drops, lance drops, knight drops, and uchifuzume conditions.
- Enhanced test coverage for edge cases involving checkmate scenarios and piece interactions, ensuring robust legality checks for pawn drops.
