+++ packages/rust-core/crates/engine-core/src/movegen/mod.rs
@@ -1,9 +1,7 @@
 mod compat;
 pub mod generator;
-pub mod move_picker;
 
 #[cfg(test)]
 mod tests;
 
 pub use compat::MoveGen;
-pub use move_picker::MovePicker;
