--- packages/rust-core/crates/engine-core/src/movegen/move_picker/tests/capture_tests.rs
@@ -1,216 +1,0 @@
-//! Capture-related tests
-
-use crate::movegen::move_picker::move_scoring::{CAPTURE_PROMO_TIE_BREAK, SEE_PACK_SHIFT};
-use crate::movegen::move_picker::MovePicker;
-use crate::search::types::SearchStack;
-use crate::shogi::{Move, Position};
-use crate::usi::{parse_usi_move, parse_usi_square};
-use crate::History;
-
-#[test]
-fn test_see_calculation() {
-    // Create a position where we can test SEE
-    let mut pos = Position::startpos();
-
-    // Create a position with some captures
-    let setup_moves = [
-        parse_usi_move("7g7f").unwrap(), // 先手の歩
-        parse_usi_move("3c3d").unwrap(), // 後手の歩
-        parse_usi_move("2g2f").unwrap(), // 先手の歩
-        parse_usi_move("3d3e").unwrap(), // 後手の歩
-        parse_usi_move("2f2e").unwrap(), // 先手の歩
-        parse_usi_move("3e3f").unwrap(), // 後手の歩
-    ];
-
-    for mv in &setup_moves {
-        pos.do_move(*mv);
-    }
-
-    let history = History::new();
-    let stack = SearchStack::default();
-    let picker = MovePicker::new(&pos, None, None, &history, &stack, 1);
-
-    // Test capturing the pawn at 3f with our pawn at 3g
-    // This should be a good capture (pawn for pawn)
-    let capture_3f = parse_usi_move("3g3f").unwrap(); // 先手の歩が後手の歩を取る
-    let see_value = picker.see(capture_3f);
-    assert_eq!(see_value, 100, "Pawn x Pawn should have SEE value of 100 (pawn value)");
-
-    // If there were a more valuable piece defending, SEE would be negative
-    // But in this simple position, it's just pawn for pawn
-}
-
-#[test]
-fn test_mvv_lva_ordering() {
-    // MVV-LVA ordering is now handled by SEE
-    // This test verifies that captures are ordered correctly
-    let mut pos = Position::startpos();
-
-    // Make some moves to create capture opportunities
-    let moves = [
-        Move::normal(parse_usi_square("7g").unwrap(), parse_usi_square("7f").unwrap(), false), // Black pawn forward
-        Move::normal(parse_usi_square("6c").unwrap(), parse_usi_square("6d").unwrap(), false), // White pawn forward
-        Move::normal(parse_usi_square("7f").unwrap(), parse_usi_square("7e").unwrap(), false), // Black pawn forward
-        Move::normal(parse_usi_square("6d").unwrap(), parse_usi_square("6e").unwrap(), false), // White pawn forward
-    ];
-
-    for mv in &moves {
-        pos.do_move(*mv);
-    }
-
-    let history = History::new();
-    let stack = SearchStack::default();
-    let mut picker = MovePicker::new(&pos, None, None, &history, &stack, 1);
-
-    // Collect captures
-    let mut captures = Vec::new();
-    while let Some(mv) = picker.next_move() {
-        if !mv.is_drop() && pos.board.piece_on(mv.to()).is_some() {
-            captures.push(mv);
-        }
-        if captures.len() >= 5 {
-            break;
-        }
-    }
-
-    // Should have some captures
-    assert!(!captures.is_empty(), "Should have found some captures");
-}
-
-#[test]
-fn test_see_negative_promoting_capture() {
-    // Test that SEE-negative promoting captures are correctly classified as bad captures
-    // This tests the fix for the bug where promotion bonus could make bad captures appear good
-
-    // Create a position where a silver can capture a defended pawn with promotion
-    // Use startpos and make specific moves to set up the test scenario
-    let mut pos = Position::startpos();
-
-    // Setup moves to create the test position
-    let setup_moves = [
-        parse_usi_move("7g7f").unwrap(),  // Black
-        parse_usi_move("3c3d").unwrap(),  // White
-        parse_usi_move("6g6f").unwrap(),  // Black
-        parse_usi_move("3a3b").unwrap(),  // White (silver defends)
-        parse_usi_move("6f6e").unwrap(),  // Black
-        parse_usi_move("8c8d").unwrap(),  // White
-        parse_usi_move("6e6d").unwrap(),  // Black (silver advances)
-        parse_usi_move("6c6d").unwrap(),  // White (pawn blocks)
-        parse_usi_move("7i6h").unwrap(),  // Black (silver up)
-        parse_usi_move("8d8e").unwrap(),  // White
-        parse_usi_move("6h6g").unwrap(),  // Black
-        parse_usi_move("8e8f").unwrap(),  // White
-        parse_usi_move("6g5f").unwrap(),  // Black
-        parse_usi_move("8f8g+").unwrap(), // White
-        parse_usi_move("5f4e").unwrap(),  // Black (silver advances)
-        parse_usi_move("8g8f").unwrap(),  // White
-        parse_usi_move("4e3d").unwrap(),  // Black (silver to 3d)
-    ];
-
-    for mv in &setup_moves {
-        pos.do_move(*mv);
-    }
-
-    let history = History::new();
-    let stack = SearchStack::default();
-    let mut picker = MovePicker::new(&pos, None, None, &history, &stack, 1);
-
-    // The move 3d2c+ (silver capturing pawn with promotion) should have negative SEE
-    // because white silver on 3b can recapture
-    let promoting_capture = Move::normal(
-        parse_usi_square("3d").unwrap(),
-        parse_usi_square("2c").unwrap(),
-        true, // promote
-    );
-
-    // Check if the move is actually a capture in this position
-    if picker.get_captured_piece(promoting_capture).is_none() {
-        // If not a capture, try a simpler test with a general principle
-        // Generate all captures and check any promoting capture with negative SEE
-        picker.generate_captures();
-        picker.score_captures();
-
-        let mut found_test = false;
-        for scored_move in &picker.moves {
-            if scored_move.mv.is_promote() && picker.get_captured_piece(scored_move.mv).is_some() {
-                let see = picker.see(scored_move.mv);
-                if see < 0 {
-                    found_test = true;
-                    // The key test: packed score must preserve negative sign
-                    assert!(
-                        scored_move.score < 0,
-                        "Packed score for SEE-negative promoting capture must be negative. \
-                         Move: {:?}, SEE: {}, Packed score: {}",
-                        scored_move.mv,
-                        see,
-                        scored_move.score
-                    );
-
-                    // Additional check: verify the packed score format
-                    // Upper bits should contain SEE value
-                    let packed_see = scored_move.score >> SEE_PACK_SHIFT;
-                    assert_eq!(
-                        packed_see, see,
-                        "Upper bits should contain SEE value. Expected: {see}, Actual: {packed_see}"
-                    );
-
-                    // Lower bits should contain tie-break value
-                    let tie_break_mask = (1 << SEE_PACK_SHIFT) - 1;
-                    let packed_tie_break = scored_move.score & tie_break_mask;
-                    assert_eq!(
-                        packed_tie_break,
-                        CAPTURE_PROMO_TIE_BREAK,
-                        "Lower bits should contain promotion tie-break. Expected: {CAPTURE_PROMO_TIE_BREAK}, Actual: {packed_tie_break}"
-                    );
-                    break;
-                }
-            }
-        }
-
-        if !found_test {
-            // If no negative SEE promoting captures exist, test passed
-            // The important thing is the logic is correct when they do exist
-        }
-    } else {
-        // Verify this specific move has negative SEE
-        let see_value = picker.see(promoting_capture);
-
-        // Generate and score all captures
-        picker.generate_captures();
-        picker.score_captures();
-
-        // Find our specific move in the generated captures
-        let found_move = picker.moves.iter().find(|sm| sm.mv == promoting_capture);
-
-        // Ensure the move was generated and has the correct score
-        if let Some(scored_move) = found_move {
-            if see_value < 0 {
-                // The key test: packed score must preserve negative sign
-                assert!(
-                    scored_move.score < 0,
-                    "Packed score for SEE-negative promoting capture must be negative. \
-                     Move: {:?}, SEE: {see_value}, Packed score: {}",
-                    scored_move.mv,
-                    scored_move.score
-                );
-
-                // Additional check: verify the packed score format
-                // Upper bits should contain SEE value
-                let packed_see = scored_move.score >> SEE_PACK_SHIFT;
-                assert_eq!(
-                    packed_see, see_value,
-                    "Upper bits should contain SEE value. Expected: {see_value}, Actual: {packed_see}"
-                );
-
-                // Lower bits should contain tie-break value
-                let tie_break_mask = (1 << SEE_PACK_SHIFT) - 1;
-                let packed_tie_break = scored_move.score & tie_break_mask;
-                assert_eq!(
-                    packed_tie_break,
-                    CAPTURE_PROMO_TIE_BREAK,
-                    "Lower bits should contain promotion tie-break. Expected: {CAPTURE_PROMO_TIE_BREAK}, Actual: {packed_tie_break}"
-                );
-            }
-        }
-    }
-}
