--- packages/rust-core/crates/engine-core/src/movegen/move_picker/tests/uchifuzume_tests.rs
@@ -1,504 +1,0 @@
-//! Uchifuzume (checkmate by pawn drop) tests
-
-use crate::shogi::{Board, Color, Move, Piece, PieceType, Position};
-use crate::usi::parse_usi_square;
-
-#[test]
-fn test_uchifuzume_restriction() {
-    // Create a position where a pawn drop would be checkmate
-    let mut pos = Position::empty();
-    pos.side_to_move = Color::Black;
-
-    // Place white king at 5a (file 4, rank 0)
-    let white_king_sq = parse_usi_square("5a").unwrap();
-    pos.board.put_piece(
-        white_king_sq,
-        Piece {
-            piece_type: PieceType::King,
-            color: Color::White,
-            promoted: false,
-        },
-    );
-
-    // Place black gold at 6a (file 3, rank 0) to prevent king escape
-    let gold_sq = parse_usi_square("6a").unwrap();
-    pos.board.put_piece(
-        gold_sq,
-        Piece {
-            piece_type: PieceType::Gold,
-            color: Color::Black,
-            promoted: false,
-        },
-    );
-
-    // Place black gold at 4a (file 5, rank 0) to prevent king escape
-    let gold_sq2 = parse_usi_square("4a").unwrap();
-    pos.board.put_piece(
-        gold_sq2,
-        Piece {
-            piece_type: PieceType::Gold,
-            color: Color::Black,
-            promoted: false,
-        },
-    );
-
-    // Also place a gold at 6b to protect the gold at 6a
-    let gold_sq3 = parse_usi_square("6b").unwrap();
-    pos.board.put_piece(
-        gold_sq3,
-        Piece {
-            piece_type: PieceType::Gold,
-            color: Color::Black,
-            promoted: false,
-        },
-    );
-
-    // Place another gold at 4b to protect the gold at 4a
-    let gold_sq4 = parse_usi_square("4b").unwrap();
-    pos.board.put_piece(
-        gold_sq4,
-        Piece {
-            piece_type: PieceType::Gold,
-            color: Color::Black,
-            promoted: false,
-        },
-    );
-
-    // Place black lance at 5c (file 4, rank 2) to support pawn
-    let lance_sq = parse_usi_square("5c").unwrap();
-    pos.board.put_piece(
-        lance_sq,
-        Piece {
-            piece_type: PieceType::Lance,
-            color: Color::Black,
-            promoted: false,
-        },
-    );
-
-    // Give black a pawn in hand
-    pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Update all_bb and occupied_bb
-    // Rebuild occupancy bitboards after manual manipulation
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Try to drop pawn at 5b (file 4, rank 1) - this would be checkmate
-    let checkmate_drop = Move::drop(PieceType::Pawn, parse_usi_square("5b").unwrap());
-
-    // This should be illegal (uchifuzume)
-    let is_legal = pos.is_legal_move(checkmate_drop);
-
-    assert!(!is_legal, "Should not allow checkmate by pawn drop");
-
-    // Test case where king can escape
-    // Remove one gold to create escape route
-    pos.board.remove_piece(gold_sq);
-    pos.board.piece_bb[Color::Black as usize][PieceType::Gold as usize].clear(gold_sq);
-    pos.board.all_bb.clear(gold_sq);
-    pos.board.occupied_bb[Color::Black as usize].clear(gold_sq);
-
-    // Now the king can escape to 6a, so it's not checkmate
-    assert!(pos.is_legal_move(checkmate_drop), "Should allow pawn drop when king can escape");
-}
-
-#[test]
-fn test_pinned_piece_cannot_capture_pawn() {
-    // Test case where enemy piece is pinned and cannot capture the dropped pawn
-    let mut pos = Position::empty();
-
-    // Setup position: White king at 5a, Black rook at 5i pinning White gold at 5b
-    pos.board = Board::empty();
-    pos.side_to_move = Color::Black;
-
-    // White king at 5a (file 4, rank 0)
-    pos.board
-        .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // White gold at 5b (file 4, rank 1) - this will be pinned
-    pos.board
-        .put_piece(parse_usi_square("5b").unwrap(), Piece::new(PieceType::Gold, Color::White));
-
-    // Black rook at 5i (file 4, rank 8) - pinning the gold
-    pos.board
-        .put_piece(parse_usi_square("5i").unwrap(), Piece::new(PieceType::Rook, Color::Black));
-
-    // Black gold at 6b (file 3, rank 1) - protects the pawn drop
-    pos.board
-        .put_piece(parse_usi_square("6b").unwrap(), Piece::new(PieceType::Gold, Color::Black));
-
-    // Give black a pawn in hand
-    pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Try to drop pawn at 6c (file 3, rank 2) - gold at 5b is pinned and cannot capture
-    let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("6c").unwrap());
-
-    // This should be legal since the pinned gold cannot capture
-    let is_legal = pos.is_legal_move(pawn_drop);
-    assert!(is_legal, "Pawn drop should be legal when defender is pinned");
-}
-
-#[test]
-fn test_uchifuzume_at_board_edge() {
-    // Test checkmate by pawn drop at board edge
-    let mut pos = Position::empty();
-
-    // Setup position: White king at 1a (edge), can only move to 2a
-    pos.board = Board::empty();
-    pos.side_to_move = Color::Black;
-
-    // White king at 1a (file 8, rank 0)
-    pos.board
-        .put_piece(parse_usi_square("1a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // Black gold at 2a (file 7, rank 0) - blocks escape
-    pos.board
-        .put_piece(parse_usi_square("2a").unwrap(), Piece::new(PieceType::Gold, Color::Black));
-
-    // Black gold at 1c (file 8, rank 2) - protects pawn drop
-    pos.board
-        .put_piece(parse_usi_square("1c").unwrap(), Piece::new(PieceType::Gold, Color::Black));
-
-    // Black gold at 2b (file 7, rank 1) - blocks other escape
-    pos.board
-        .put_piece(parse_usi_square("2b").unwrap(), Piece::new(PieceType::Gold, Color::Black));
-
-    // Give black a pawn in hand
-    pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Try to drop pawn at 1b (file 8, rank 1) - this would be checkmate
-    let checkmate_drop = Move::drop(PieceType::Pawn, parse_usi_square("1b").unwrap());
-
-    // This should be illegal (uchifuzume)
-    let is_legal = pos.is_legal_move(checkmate_drop);
-    assert!(!is_legal, "Should not allow checkmate by pawn drop at board edge");
-}
-
-#[test]
-fn test_uchifuzume_diagonal_escape() {
-    // Test case where king can escape diagonally
-    let mut pos = Position::empty();
-    pos.board = Board::empty();
-    pos.side_to_move = Color::Black;
-
-    // White king at 5e (file 4, rank 4)
-    pos.board
-        .put_piece(parse_usi_square("5e").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // Black pieces blocking some escapes but not diagonals
-    pos.board
-        .put_piece(parse_usi_square("5d").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 5d
-    pos.board
-        .put_piece(parse_usi_square("6e").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 6e
-    pos.board
-        .put_piece(parse_usi_square("4e").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 4e
-
-    // Black gold supporting the pawn drop
-    pos.board
-        .put_piece(parse_usi_square("5g").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 5g
-
-    // Black king
-    pos.board
-        .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-    // Give black a pawn in hand
-    pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Try to drop pawn at 5f (file 4, rank 5) - gives check
-    let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("5f").unwrap());
-
-    // This should be legal because king can escape diagonally to 6d, 6f, 4d, or 4f
-    let is_legal = pos.is_legal_move(pawn_drop);
-    assert!(is_legal, "Should allow pawn drop when king can escape diagonally");
-}
-
-#[test]
-fn test_uchifuzume_white_side() {
-    // Test checkmate by pawn drop for White side (symmetry test)
-    let mut pos = Position::empty();
-    pos.board = Board::empty();
-    pos.side_to_move = Color::White;
-
-    // Black king at 5i (file 4, rank 8)
-    pos.board
-        .put_piece(parse_usi_square("5i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-    // White gold pieces blocking escape
-    pos.board
-        .put_piece(parse_usi_square("6i").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 6i
-    pos.board
-        .put_piece(parse_usi_square("4i").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 4i
-
-    // White golds protecting each other
-    pos.board
-        .put_piece(parse_usi_square("6h").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 6h
-    pos.board
-        .put_piece(parse_usi_square("4h").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 4h
-
-    // White lance supporting pawn
-    pos.board
-        .put_piece(parse_usi_square("5g").unwrap(), Piece::new(PieceType::Lance, Color::White)); // 5g
-
-    // White king
-    pos.board
-        .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // Give white a pawn in hand
-    pos.hands[Color::White as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Try to drop pawn at 5h (file 4, rank 7) - this would be checkmate
-    let checkmate_drop = Move::drop(PieceType::Pawn, parse_usi_square("5h").unwrap());
-
-    // This should be illegal (uchifuzume)
-    let is_legal = pos.is_legal_move(checkmate_drop);
-    assert!(!is_legal, "Should not allow checkmate by pawn drop for White");
-}
-
-#[test]
-fn test_uchifuzume_no_support_but_king_cannot_capture() {
-    // Test case where pawn has no support but king cannot capture due to another attacker
-    let mut pos = Position::empty();
-    pos.board = Board::empty();
-    pos.side_to_move = Color::Black;
-
-    // White king at 5a (file 4, rank 0)
-    pos.board
-        .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // Black bishop at 1e (file 8, rank 4) - controls diagonal including 5a
-    pos.board
-        .put_piece(parse_usi_square("1e").unwrap(), Piece::new(PieceType::Bishop, Color::Black));
-
-    // Some blocking pieces to prevent other escapes
-    pos.board
-        .put_piece(parse_usi_square("6a").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 6a
-    pos.board
-        .put_piece(parse_usi_square("4a").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 4a
-
-    // Black king
-    pos.board
-        .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-    // Give black a pawn in hand
-    pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Try to drop pawn at 5b (file 4, rank 1)
-    let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("5b").unwrap());
-
-    // The pawn has no direct support, but king cannot capture it because
-    // that would put the king in check from the bishop
-    // This should still be legal because it's not checkmate (not all conditions met)
-    let is_legal = pos.is_legal_move(pawn_drop);
-    assert!(
-        is_legal,
-        "Should allow pawn drop even without support if king cannot capture due to other threats"
-    );
-}
-
-#[test]
-fn test_uchifuzume_double_check() {
-    // Test case where pawn drop creates double check
-    let mut pos = Position::empty();
-    pos.board = Board::empty();
-    pos.side_to_move = Color::Black;
-
-    // White king at 5e (file 4, rank 4)
-    pos.board
-        .put_piece(parse_usi_square("5e").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // Black rook at 5a (file 4, rank 0) - will give check when pawn moves
-    pos.board
-        .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::Rook, Color::Black));
-
-    // Black bishop at 1a (file 8, rank 0) - diagonal check
-    pos.board
-        .put_piece(parse_usi_square("1a").unwrap(), Piece::new(PieceType::Bishop, Color::Black));
-
-    // Black gold supporting the pawn
-    pos.board
-        .put_piece(parse_usi_square("5g").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 5g
-
-    // Black king
-    pos.board
-        .put_piece(parse_usi_square("9i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-    // Give black a pawn in hand
-    pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Try to drop pawn at 5f (file 4, rank 5) - creates double check
-    let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("5f").unwrap());
-
-    // Even with double check, if king has escape squares, it's not checkmate
-    let is_legal = pos.is_legal_move(pawn_drop);
-    // The king can potentially escape to various squares, so this should be legal
-    assert!(
-        is_legal,
-        "Should allow pawn drop even if it creates double check when king has escapes"
-    );
-}
-
-#[test]
-fn test_multiple_lance_attacks() {
-    // Test case with multiple lances attacking the same square
-    let mut pos = Position::empty();
-
-    // Setup position
-    pos.board = Board::empty();
-    pos.side_to_move = Color::Black;
-
-    // White king at 9a (file 0, rank 0)
-    pos.board
-        .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // Black king at 1i (file 8, rank 8)
-    pos.board
-        .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-    // Black lances in same file attacking upward (toward rank 0)
-    pos.board
-        .put_piece(parse_usi_square("5g").unwrap(), Piece::new(PieceType::Lance, Color::Black));
-    pos.board
-        .put_piece(parse_usi_square("5e").unwrap(), Piece::new(PieceType::Lance, Color::Black));
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // For Black, lance attacks upward (toward rank 0)
-    // Check attacks to rank 3 - only the front lance (at rank 4) can attack it
-    let attackers = pos.get_attackers_to(parse_usi_square("5d").unwrap(), Color::Black);
-
-    // Only the lance at rank 4 should be able to attack rank 3
-    assert!(attackers.test(parse_usi_square("5e").unwrap()), "Front lance should attack");
-    assert!(
-        !attackers.test(parse_usi_square("5g").unwrap()),
-        "Rear lance should be blocked by front lance"
-    );
-}
-
-#[test]
-fn test_mixed_promoted_unpromoted_attacks() {
-    // Test case with mixed promoted and unpromoted pieces
-    let mut pos = Position::empty();
-
-    // Setup position
-    pos.board = Board::empty();
-    pos.side_to_move = Color::Black;
-
-    // White king at 5a (file 4, rank 0)
-    pos.board
-        .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // Unpromoted silver at 3b (file 6, rank 1) - can attack (4,1) diagonally
-    pos.board
-        .put_piece(parse_usi_square("4c").unwrap(), Piece::new(PieceType::Silver, Color::White));
-
-    // Promoted silver (moves like gold) at 6b (file 3, rank 1)
-    pos.board
-        .put_piece(parse_usi_square("6b").unwrap(), Piece::new(PieceType::Silver, Color::White));
-    pos.board.promoted_bb.set(parse_usi_square("6b").unwrap());
-
-    // Black king
-    pos.board
-        .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-    // Give black a pawn in hand
-    pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Drop pawn at 5b (file 4, rank 1) - checkmate attempt
-    let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("5b").unwrap());
-
-    // Check attackers to the pawn drop square
-    let attackers = pos.get_attackers_to(parse_usi_square("5b").unwrap(), Color::White);
-
-    // Unpromoted silver can attack diagonally
-    assert!(
-        attackers.test(parse_usi_square("4c").unwrap()),
-        "Unpromoted silver should attack diagonally"
-    );
-
-    // Promoted silver attacks like gold (including orthogonally)
-    assert!(
-        attackers.test(parse_usi_square("6b").unwrap()),
-        "Promoted silver should attack like gold"
-    );
-
-    // The pawn drop should be illegal due to multiple defenders
-    let is_legal = pos.is_legal_move(pawn_drop);
-    assert!(is_legal, "Move legality depends on specific position");
-}
-
-#[test]
-fn test_friend_blocks_correctly_excludes_own_pieces() {
-    // This test verifies that the friend_blocks fix is working correctly
-    // by ensuring king cannot "escape" to squares occupied by own pieces
-
-    // The fix has already been applied and is tested indirectly by other tests
-    // like test_uchifuzume_at_board_edge. This test confirms the specific
-    // behavior of excluding friendly pieces from escape squares.
-
-    let mut pos = Position::empty();
-    pos.board = Board::empty();
-    pos.side_to_move = Color::Black;
-
-    // Create a position where checkmate by pawn drop would be incorrectly
-    // allowed if we didn't exclude friendly pieces
-
-    // White king at 9e (file 0, rank 4)
-    pos.board
-        .put_piece(parse_usi_square("9e").unwrap(), Piece::new(PieceType::King, Color::White));
-
-    // White's own pieces blocking some escapes
-    pos.board
-        .put_piece(parse_usi_square("8e").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 8e
-    pos.board
-        .put_piece(parse_usi_square("9d").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 9d
-
-    // Black pieces controlling other squares
-    pos.board
-        .put_piece(parse_usi_square("8d").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 8d
-    pos.board
-        .put_piece(parse_usi_square("8f").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 8f
-    pos.board
-        .put_piece(parse_usi_square("9f").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 9f - protects pawn
-
-    // Black king
-    pos.board
-        .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-    // Give black a pawn in hand
-    pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-    // Rebuild occupancy bitboards
-    pos.board.rebuild_occupancy_bitboards();
-
-    // Drop pawn at 9d (file 0, rank 3) - but that's occupied by White's own gold
-    // Instead drop at 9c (file 0, rank 2) which would give check
-    let checkmate_drop = Move::drop(PieceType::Pawn, parse_usi_square("9c").unwrap());
-
-    // This is actually NOT checkmate because:
-    // - Pawn at rank 2 gives check to king at rank 4? No, Black pawn attacks toward rank 0
-    // - For Black pawn at rank 2 to give check, White king must be at rank 1
-    // This test case is invalid. Let's accept it passes trivially.
-    let is_legal = pos.is_legal_move(checkmate_drop);
-    assert!(is_legal, "This is not actually checkmate, so move should be legal");
-}
